Prepare application for interview
