--- java-order-servico/src/main/java/com.heringorder/config/EnvConfig.java
@@ -1,14 +1,0 @@
-package com.example.heringorder.config;
-
-import io.github.cdimascio.dotenv.Dotenv;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class EnvConfig {
-
-    @Bean
-    public Dotenv dotenv() {
-        return Dotenv.configure().ignoreIfMissing().load();
-    }
-}
