--- java-order-servico/src/main/java/com.heringorder/models/Order.java
@@ -1,19 +1,0 @@
-package com.example.heringorder.models;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "orders")
-@Data
-@NoArgsConstructor
-public class Order {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-  private Long userId;
-  private Double totalAmount;
-  private String status; 
-}
