--- java-order-servico/src/main/java/com.heringorder/resources/application.properties
@@ -1,9 +1,0 @@
-server.port=${SERVER_PORT:8081}
-
-spring.datasource.url=jdbc:postgresql://${DB_HOST:${DB_HOST}}:${DB_PORT:${DB_PORT}}/${DB_NAME:${DB_NAME}}
-spring.datasource.username=${DB_USER:${DB_USER}}
-spring.datasource.password=${DB_PASS:${DB_PASS}}
-
-# JPA
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
