+++ java-order-servico/src/main/java/com/heringorder/HeringOrderServiceApplication.java
@@ -1,0 +1,11 @@
+package com.heringorder;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HeringOrderServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(HeringOrderServiceApplication.class, args);
+    }
+}
