+++ java-order-servico/src/main/java/com/heringorder/models/Order.java
@@ -1,0 +1,18 @@
+package com.heringorder.models;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import jakarta.persistence.*;
+
+@Data
+@NoArgsConstructor
+@Entity
+public class Order {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long userId;
+    private Double totalAmount;
+    private String status;
+}
