+++ java-order-servico/src/main/java/com/heringorder/repositories/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.heringorder.repositories;
+
+import com.heringorder.models.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
