+++ java-order-servico/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+server.port=8080
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/hering_order_db
+spring.datasource.username=hering_user
+spring.datasource.password=minhasenha
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
